:recycle: moved files do dirs
