+++ aksel.nav.no/website/components/layout/header/search/hooks/index.ts
@@ -1,0 +1,3 @@
+export * from "./useDebounce";
+export * from "./useSearch";
+export * from "./useShortcut";
