+++ aksel.nav.no/website/components/layout/header/search/hooks/useShortcut.ts
