+++ aksel.nav.no/website/components/layout/header/search/utils/format-result.ts
