+++ aksel.nav.no/website/components/layout/header/search/utils/group-results.ts
