+++ aksel.nav.no/website/components/layout/header/search/utils/index.ts
@@ -1,0 +1,3 @@
+export * from "./format-result";
+export * from "./fuse-search";
+export * from "./group-results";
